ctp_pr0: add data filter for magnetic sensor
BZ:38778

This patch add a filter for magnetic sensor
to remove magnetic noise from other devices
on the platform.
Filter employs a 40 length buffer to buff
most recent data. When a new data comes, the
average of the buffed data and new data are
calculated as the output data.
This approach will remove noise but also will
delay the sensor's response when magnetic field
change fast.

Change-Id: I916bffe21616c1d35ba0523f8d26b35ff3b02bb7
Signed-off-by: jli127 <jian.d.li@intel.com>
Reviewed-on: http://android.intel.com:8080/49117
Reviewed-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: Liu, Hong <hong.liu@intel.com>
Reviewed-by: Du, Alek <alek.du@intel.com>
Reviewed-by: Tang, HaifengX <haifengx.tang@intel.com>
Tested-by: Wang, Zhifeng <zhifeng.wang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>